Loop and test restructuring (#9383)
