+++ pyproject.toml
@@ -62,10 +62,8 @@ ignore_errors = "True"
 [[tool.mypy.overrides]]
 module = [
     "pytorch_lightning.callbacks.pruning",
-    "pytorch_lightning.loops.closure",
-    "pytorch_lightning.loops.batch.manual",
-    "pytorch_lightning.loops.optimizer",
-    "pytorch_lightning.trainer.evaluation_loop",
+    "pytorch_lightning.loops.optimization.*",
+    "pytorch_lightning.loops.evaluation_loop",
     "pytorch_lightning.trainer.connectors.logger_connector.*",
     "pytorch_lightning.trainer.progress",
     "pytorch_lightning.tuner.auto_gpu_select",
