+++ pytorch_lightning/loops/__init__.py
@@ -18,4 +18,4 @@ from pytorch_lightning.loops.batch import TrainingBatchLoop  # noqa: F401
 from pytorch_lightning.loops.dataloader import DataLoaderLoop, EvaluationLoop, PredictionLoop  # noqa: F401
 from pytorch_lightning.loops.epoch import EvaluationEpochLoop, PredictionEpochLoop, TrainingEpochLoop  # noqa: F401
 from pytorch_lightning.loops.fit_loop import FitLoop  # noqa: F401
-from pytorch_lightning.loops.optimizer.optimizer_loop import OptimizerLoop  # noqa: F401
+from pytorch_lightning.loops.optimization.optimizer_loop import OptimizerLoop  # noqa: F401
