+++ pytorch_lightning/loops/batch/__init__.py
@@ -12,5 +12,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from pytorch_lightning.loops.batch.manual import ManualOptimization  # noqa: F401
 from pytorch_lightning.loops.batch.training_batch_loop import TrainingBatchLoop  # noqa: F401
+from pytorch_lightning.loops.optimization.manual_loop import ManualOptimization  # noqa: F401
