+++ pytorch_lightning/loops/optimization/__init__.py
@@ -12,4 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from pytorch_lightning.loops.optimizer.optimizer_loop import OptimizerLoop  # noqa: F401
+from pytorch_lightning.loops.optimization.optimizer_loop import OptimizerLoop  # noqa: F401
