+++ pytorch_lightning/loops/optimization/manual_loop.py
@@ -14,7 +14,7 @@
 from typing import Any, Optional
 
 from pytorch_lightning.loops import Loop
-from pytorch_lightning.loops.closure import ClosureResult
+from pytorch_lightning.loops.optimization.closure import ClosureResult
 from pytorch_lightning.loops.utilities import (
     _build_training_step_kwargs,
     _check_training_step_output,
