+++ tests/loops/test_all.py
